feat(GAME-060): gen-assets — unified asset generation tool (SVG sprites + PNG images)
Replaces gen-sprites.main.kts + gen-images.test.main.kts with a single
gen-assets.main.kts (57 tests) supporting two subcommands:

  gen-assets sprites   — SVG via Gemini/Grok chat API (was gen-sprites)
  gen-assets images    — PNG via Gemini Imagen / Grok image API (new)

Verified working image models (2026-05-04):
  Gemini: imagen-4.0-generate-001
  Grok:   grok-imagine-image, grok-imagine-image-pro

Shared across subcommands: credential resolution, spec loading,
Moshi/OkHttp infrastructure, all CLI flags.
